Documents/GitHub/ai-bi-genie/requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
altair==5.2.0

# Database connectors
pyodbc==5.0.1
pymssql==2.2.11
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# Azure SDK
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-identity==1.15.0
azure-cosmos==4.5.1
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-mgmt-sql==4.0.0
azure-ai-textanalytics==5.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-monitor-query==1.2.0

# Data processing
dask==2023.12.1
polars==0.20.3
pyarrow==14.0.2
fastparquet==2023.10.1

# Machine Learning
scikit-learn==1.4.0
xgboost==2.0.3
lightgbm==4.2.0
prophet==1.1.5
statsmodels==0.14.1

# Deep Learning (optional, for advanced analytics)
torch==2.1.2
transformers==4.36.2

# API and web frameworks
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
httpx==0.26.0

# Data visualization
matplotlib==3.8.2
seaborn==0.13.1
bokeh==3.3.3
dash==2.14.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
typer==0.9.0
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pre-commit==3.6.0

# Caching and performance
redis==5.0.1
diskcache==5.6.3
joblib==1.3.2

# File handling
openpyxl==3.1.2
xlsxwriter==3.1.9
python-docx==1.1.0
pypdf==3.17.4
python-pptx==0.6.23

# Scheduling and background tasks
celery==5.3.4
apscheduler==3.10.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring and observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0

# Data quality and validation
great-expectations==0.18.8
pandera==0.17.2

# Natural Language Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Time series analysis
tsfresh==0.20.2
sktime==0.25.0

# Graph analytics (optional)
networkx==3.2.1
igraph==0.11.3

# Geospatial (optional)
geopandas==0.14.2
folium==0.15.1

# Database migrations
alembic==1.13.1

# Environment management
python-decouple==3.8

# API documentation
mkdocs==1.5.3
mkdocs-material==9.5.3
mkdocstrings==0.24.0

# Performance profiling
memory-profiler==0.61.0
line-profiler==4.1.2

# Distributed computing (optional)
ray==2.9.0
dask-ml==2023.3.24

# Business Intelligence specific
pygwalker==0.4.7
dtale==3.9.0

# Email and notifications
sendgrid==6.11.0
twilio==8.11.0

# Feature store (optional)
feast==0.35.0

# Model registry and MLOps
mlflow==2.9.2
evidently==0.4.13

# Data versioning
dvc==3.39.0

# Kafka integration (for real-time data)
kafka-python==2.0.2
confluent-kafka==2.3.0

# Elasticsearch integration
elasticsearch==8.11.1

# GraphQL (optional)
strawberry-graphql==0.215.1

# WebSocket support
websockets==12.0
python-socketio==5.10.0

# Rate limiting
slowapi==0.1.9

# HTML parsing and web scraping (if needed)
beautifulsoup4==4.12.2
requests-html==0.10.0

# Jupyter integration (for notebook support)
jupyter==1.0.0
jupyterlab==4.0.10
ipywidgets==8.1.1

# Code quality and security
bandit==1.7.6
safety==3.0.1

# Documentation generation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Deployment and packaging
gunicorn==21.2.0
wheel==0.42.0
setuptools==69.0.3

# Cloud-native tools
kubernetes==29.0.0
docker==7.0.0

# Configuration management
hydra-core==1.3.2
omegaconf==2.3.0

# Data contracts
dataclasses-json==0.6.3
marshmallow==3.20.2

# API rate limiting and caching
cachetools==5.3.2
aiocache==0.12.2

# Async support
aiohttp==3.9.1
aiofiles==23.2.1

# Debugging and profiling
py-spy==0.3.14
viztracer==0.16.1

# Databricks integration
databricks-sdk==0.16.0
databricks-connect==14.2.0

# Snowflake integration (optional)
snowflake-connector-python==3.6.0

# BigQuery integration (optional)
google-cloud-bigquery==3.14.1

# Version management
bump2version==1.0.1